models/AI-Model-Zoo/VAI-1.3-Model-Zoo-Code/PyTorch/pt_salsanext_semantic-kitti_64_2048_0.6_20.4G_1.3/requirements.txt
absl-py==0.10.0
astor==0.8.1
certifi==2020.6.20
cycler==0.10.0
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
importlib-metadata==2.0.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.3.1
matplotlib==3.3.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
ninja==1.10.0.post2
opencv-python==4.4.0.44
opt-einsum==3.3.0
Pillow==7.2.0
protobuf==3.13.0
pybind11==2.5.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
scikit-learn==0.23.2
scipy==1.3.1
six==1.15.0
sklearn==0.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.49.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.1
